+++ lib/openstudio/analysis/workflow.rb
@@ -124,6 +124,10 @@ module OpenStudio
         end
       end
 
+      # Iterate over all the WorkflowItems
+      def each
+        @items.each { |i| yield i }
+      end
       # Find the measure by its instance name
       #
       # @params instance_name [String] instance name of the measure