+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Version 0.4.2 (Unreleased)
 -------------
 * Bug fix when adding measure from path, this now sets the correct argument name.
 * Fix namespace conflict with OpenStudio::Logger and OpenStudio::Time
+* Create method for saving the Analysis Zip file (save_analysis_zip)
+
 Version 0.4.1
 -------------
 * Bug fix to address the spec/files directory being prepended to the measures